Remove MachineUpdater
